meetingreservitionroom/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-filter==22.1
django-jalali-date==1.0.1
jdatetime==4.1.0
numpy==1.23.5
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
psycopg2-binary